+++ scripts/postinstall.js
@@ -1,12 +1,12 @@
 'use strict';
 
-const dapServerVersion = '1.7.0';
+const dapServerVersion = '1.8.0-SNAPSHOT';
 
 const download = require('mvn-artifact-download').default;
 const fs = require('fs');
 const path = require('path');
 
-const MAVEN_REPO_URL = 'https://oss.sonatype.org/content/repositories/releases/';
+const MAVEN_REPO_URL = 'https://oss.sonatype.org/content/repositories/snapshots/';
 
 download('com.github.camel-tooling:camel-dap-server:' + dapServerVersion, './jars/', MAVEN_REPO_URL).then((filename)=>{
 	fs.renameSync(filename, path.join('.', 'jars', 'camel-dap-server.jar'));