+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the "vscode-debug-adapter-apache-camel" extension will be
 
 ## 1.7.0
 
-
+- Update default Camel version used for Camel JBang from 4.9.0 to 4.10.2
 
 ## 1.6.0